parser/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
dnspython==2.7.0
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
openai==1.82.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic-core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing-extensions==4.13.2
typing-inspection==0.4.1
urllib3==2.5.0
webencodings==0.5.1
